Move UtilsTablePropertiesTest to configuration module
